Make meta buildpacks reference released version images
Pointing to unreleased code results in different versions being packaged between meta-buildpacks
and standalone buildpacks. This is confusing to users and pack itself. This weakens the testing
strategy a bit since we cannot test a meta-buildpack with unreleased changes of a sub buildpack.
This will be addressed in a later commit.