Fix Dependabot configuration for Cargo
Since Dependabot is looking in the repo root, and that only works if either
the crate exists there, or the repo is using cargo workspaces with a suitably
configured repo root `Cargo.toml`.

Until this repo uses cargo workspaces (and the build scripts are adapted to work
with workspaces), the Dependabot config needs to reference each crate path explicitly.

GUS-W-9839985.