JVM invoker: Enable clippy::pedantic and unused_crate_dependencies lints
Clippy has a `pedantic` lint category for lints that are more opinionated,
or that sometimes have a higher false positive rate.

Many of these lints are actually very useful (particularly for those new to Rust,
like we are), and so we should take advantage of them - with the understanding
that it's absolutely fine to disable individual pedantic lints in the future that we
find annoying/counter-productive.

The list of all Clippy lints by category can be found here:
https://rust-lang.github.io/rust-clippy/stable/index.html

In addition, the `unused_crate_dependencies` lint has been enabled to ensure the
crate doesn't have unused dependencies:
https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#unused-crate-dependencies

This is a port of:
https://github.com/heroku/heroku-buildpacks-repo-template/pull/6
https://github.com/heroku/heroku-buildpacks-repo-template/pull/7
https://github.com/heroku/heroku-buildpacks-repo-template/pull/8

GUS-W-9888451.
GUS-W-9888470.