Failing test for logger
It was reported that no logs are coming out of the logger interface for the java function. We currently have no tests asserting log output is being emitted from functions. This commit exercises a logger and adds a (failing) assertion that the contents of that logger is then emitted from the container logs.

GUS-W-9933414