Ignore whitespace in release script
Currently if you attempt to release a buildpack with id `heroku/java   ` (with whitespace before or after) it will fail. It also takes a few minutes to fail and isn't 100% obvious in the message that the failure is due to whitespace.

This trims whitespace from the input on both release scripts. I also modified the error message to quote the variable so if it's used in the future without being trimmed first, the error would be easier to detect.


https://github.com/heroku/buildpacks-nodejs/pull/132