Upgrade `heroku/procfile` to `1.0.1` (#300)
* Upgrade heroku/procfile to 1.0.1

* Add Procfile to app-with-compile-error test fixture

Older heroku/procfile versions did not fail detection if there is no Procfile, allowing tests that
explicitly add it to the list of buildpacks to pass even if there is no Procfile. heroku/procfile >= 1.0.0
will fail detection, causing these tests to fail.

To work around this, we add a Procfile to the app-with-compile-error test fixture.