maven buildpack environment configuration (#508)
* Honor buildpack provided env vars

Lifecycle sets buildpack provided build time env vars in the current
environment. This change reads from both the platform directory (user or
platform provided) as well as now the current environment (buildpack
provided). This fixes compatibility with a nexus/artifactory type
buildpack that sets `MAVEN_SETTINGS_PATH` or `MAVEN_SETTINGS_URL`.

* Recognize `JAVA_HOME` env var being set

Currently, `JAVA_HOME` is ignored even if set by a buildpack, user, or
operator since `clear-env` is set to true. This change sets it in the
maven environment if it's present.

* add changelog entry