Change requiring JVM to a use file based acceptlist
There are several high quality signals that an application needs the JVM (looking at the `sbt` and `maven` buildpacks) and we can also add more (such as checking for a `system.properties` file). Developers who want to opt into the behavior that don't have one of these common files will need to add one.

This change is breaking for developers relying on the current "always provide" logic.

## Approach

The detect logic for buildpacks (sbt and maven) are moved to the shared code crate and they can be re-used by the `heroku/jvm` buildpack. That means any app that currently uses `heroku/java` or `heroku/scala` are guaranteed to have the JVM installed by `heroku/jvm` in addition the following files are also checked:

- system.properties
- *.war // TODO: Add glob