Update buildpack metadata in buildpack.toml (#603)
Adjusts the buildpack `name`, `description` and `keywords` in the
`buildpack.toml` of all buildpacks to match the style discussed in:
https://github.com/heroku/cnb-builder-images/issues/408

The `keywords` field has intentionally been omitted from the functions
buildpacks since they are EOL and so don't want to be made more
discoverable. (The CNB registry doesn't currently ingest/display the
keywords currently, but may in the future.)

Lastly, I've cleaned up the `[metadata]` table/subtables to match the
concise style used by the Procfile and Python CNBs.

GUS-W-14121598.