Remove `heroku/procfile` from `heroku/java` and `heroku/scala`
Previously the `heroku/procfile` buildpack was included within the
`heroku/java` and `heroku/scala` composite buildpacks.

Whilst this is convenient (it saves having to add the Procfile CNB manually),
this approach has a number of issues:
- If an app uses multiple languages (eg `heroku/nodejs` and `heroku/java`),
  the Procfile CNB will end up being run multiple times, and potentially
  with different versions of the Procfile buildpack.
- It means multiple places need updating after Procfile CNB releases,
  and means the builder image can end up containing several
  different versions of the Procfile buildpack.

After this change, the `heroku/procfile` CNB is no longer included in the
`heroku/java` and `heroku/scala` composite buildpacks, and instead
will be included in the Heroku builder image order grouping (added
in the PR that releases this change into the builder).

This matches the approach already used by the Go, Python, PHP and
Ruby buildpacks.

Any app that sets a custom buildpack order in their `project.toml` (rather
than relying on the default buildpack order in the Heroku builder image),
or uses a non-Heroku builder (that does not similarly choose to add the
Procfile to the builder) will need to explicitly add the `heroku/procfile`
buildpack to the end of the buildpacks list in their `project.toml`.