Update tests to look at one source
This is a bit of an open question on the best place to print to by default. Stderr is for humans and stdout is for machines (pipes). All of our data is to the user so I picked the global output to be stderr. IDK if there's any practical problem with picking on or the other.

But it did cause all these tests to fail. This commit fixes it and opens up the conversation/question. We also have the ability to set the output location globally at the start of each buildpack.