Enable libcnb tracing (#779)
This enables `libcnb`'s `trace` feature, which means traces, spans, and
events are written to OpenTelemetry export files on disk in the ephemeral
build container, which can then be read by the Heroku build system or
when debugging locally with Pack (so long as a suitable `--volume` mount
has been configured).

See:
https://github.com/heroku/libcnb.rs/pull/723

Closes #768.
GUS-W-15148616.